flask-ecommerce/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.28.1